+++ pytest.ini
@@ -1,4 +1,4 @@
 # pytest.ini
 [pytest]
 DJANGO_SETTINGS_MODULE = backend.settings
-python_files = tests.py
+python_files = test_*.py *_tests.py
